+++ scripts/ingestion/database/vulnerability_db_init.py
@@ -1,11 +1,9 @@
 #!/usr/bin/env python
 import sqlite3
-from constants import DB_LOC
 
-conn = sqlite3.connect(DB_LOC)
-cursor = conn.cursor()
-
-def db_init():
+def db_init(location):
+        conn = sqlite3.connect(location)
+        cursor = conn.cursor()
         cursor.execute("""CREATE TABLE IF NOT EXISTS vulnerability_table(
                 source_id integer,
                 package varchar,
@@ -17,3 +15,5 @@ def db_init():
                 scope varchar,
                 releases varchar,
                 FOREIGN KEY(source_id) references source_table(source_id) )""")
+        
+        return conn,cursor
