+++ scripts/ingestion/parsers/vulnerability_parser.py
@@ -1,18 +1,35 @@
 #!/usr/bin/env python
 import requests
-import sqlite3
-from vulnerability_main import insert_records
-from constants import DB_LOC, VULNERABILITY_DATA
+from bs4 import BeautifulSoup
+from constants import NIST_URL, VULNERABILITY_DATA
+
+def parse_vulnerability_data(vulnerability_id):
+    url = f"{NIST_URL}{vulnerability_id}"
+    
+    response = requests.get(url)
+    if response.status_code == 200:
+        soup = BeautifulSoup(response.content, 'html.parser')
+        
+        # Find and extract the published date
+        published_date_element = soup.find('span', {'data-testid': 'vuln-published-on'})
+        published_date = published_date_element.text.strip() if published_date_element else None
+        
+        # Find and extract the last modified date
+        last_modified_date_element = soup.find('span', {'data-testid': 'vuln-last-modified-on'})
+        last_modified_date = last_modified_date_element.text.strip() if last_modified_date_element else None
+        
+        return published_date, last_modified_date
+    
+    return None, None
 
 # Define the main parser function
 def parser():
-    con = sqlite3.connect(DB_LOC)
     # Check if the file exists locally, if not, download it
     response = requests.get(VULNERABILITY_DATA)
 
         # Parse the data as JSON
     parsed_data = response.json()
-
+    results = []
     for name, vulnerability in parsed_data.items():
         for cve, info in vulnerability.items():
             if 'description' in info:
@@ -29,11 +46,8 @@ def parser():
                 scope="None"
             releases = info['releases']
             # print(f"{name}|{cve}|{description}|{debianbug}|{scope}|{releases}")
-            cur = con.cursor()
             # to insert records into the table
-            insert_records(cur,name,cve,description,debianbug,scope,releases) 
-            con.commit()
-    con.close()
-
-if __name__ == "__main__":
-    parser()
+            published_date, last_modified_date = parse_vulnerability_data(cve)
+            result = (name, cve, description,published_date, last_modified_date, debianbug, scope, releases)
+            results.append(result)
+    return results
