+++ scripts/vulnerability_main.py
@@ -1,36 +1,23 @@
 #!/usr/bin/env python
-import requests
-from bs4 import BeautifulSoup
 from ingestion.database.vulnerability_db_init import db_init
-from ingestion.constants import NIST_URL
+from ingestion.parsers.vulnerability_parser import parser
+from ingestion.constants import DB_LOC
 
-INSERT_RECS = '''INSERT OR IGNORE INTO vulnerability_table (source_id,package,vulnerability,description,published_date, last_modified_date,debianbug,scope,releases) VALUES (?,?,?,?,?,?,?,?,?)'''
-
-def parse_vulnerability_data(vulnerability_id):
-    url = f"{NIST_URL}{vulnerability_id}"
-    
-    response = requests.get(url)
-    if response.status_code == 200:
-        soup = BeautifulSoup(response.content, 'html.parser')
+def insert_records(results):
+    for name, cve, description, published_date, last_modified_date, debianbug, scope, releases in results:
+        cursor.execute('''select source_id,source_name from source_table where source_name ='{}' limit 1'''.format(name))
+        id = cursor.fetchone()
+        if id:
+            cursor.execute(INSERT_RECS, (id[0],name,cve,description,published_date, last_modified_date,debianbug,scope,str(releases))) 
+        else: 
+            cursor.execute(INSERT_RECS, ('Null',name,cve,description,published_date, last_modified_date,debianbug,scope,str(releases)))   
+            
+        conn.commit()
         
-        # Find and extract the published date
-        published_date_element = soup.find('span', {'data-testid': 'vuln-published-on'})
-        published_date = published_date_element.text.strip() if published_date_element else None
-        
-        # Find and extract the last modified date
-        last_modified_date_element = soup.find('span', {'data-testid': 'vuln-last-modified-on'})
-        last_modified_date = last_modified_date_element.text.strip() if last_modified_date_element else None
-        
-        return published_date, last_modified_date
-    
-    return None, None
-
-def insert_records(cur,name,cve,description,debianbug,scope,releases):
-    db_init()
-    cur.execute('''select source_id,source_name from source_table where source_name ='{}' limit 1'''.format(name))
-    id = cur.fetchone()
-    published_date, last_modified_date = parse_vulnerability_data(cve)
-    if id:
-        cur.execute(INSERT_RECS, (id[0],name,cve,description,published_date, last_modified_date,debianbug,scope,str(releases))) 
-    else: 
-        cur.execute(INSERT_RECS, ('Null',name,cve,description,published_date, last_modified_date,debianbug,scope,str(releases)))   
+ 
+if __name__ == "__main__":
+        conn,cursor = db_init(DB_LOC) # Initialize the database  
+        INSERT_RECS = '''INSERT OR IGNORE INTO vulnerability_table (source_id,package,vulnerability,description,published_date, last_modified_date,debianbug,scope,releases) VALUES (?,?,?,?,?,?,?,?,?)'''
+        results = parser() # Parse the data
+        insert_records(results)
+        conn.close()
